+++ lib/code42template.rb
@@ -1,4 +1,5 @@
 require 'code42template/version'
+require "code42template/adapters/heroku"
 require 'code42template/generators/app_generator'
 require 'code42template/actions'
 require 'code42template/app_builder'