+++ NEWS.md
@@ -44,3 +44,5 @@ UNRELEASED
 * Upgrade Ruby to 2.3.1
 * Upgrade Rails to 5.0
 * Drop quiet_assets in favor of Rails 5 native similar feature
+* Add Heroku support upon app creation
+* Add Puma as the default application server